querystring: move isHexTable to internal
PR-URL: https://github.com/nodejs/node/pull/11858
